Add support to generate reference docs for kube components
This adds Markdown generators for the following kube components:

- cloud-controller-manager
- kube-apiserver
- kube-controller-manager
- kube-proxy
- kube-scheduler
- kubelet
- kubeadm

The generator logic is based on the doc generator logic from cobra.
The main difference from the builtin markdown generator is that we are
generating HTML tables for the kube component options.
The logic is applicable to kubeadm as well.

Note that the vendored source files are removed in this PR because
they are making the doc generation dependent to the vendored version.