Update API reference generator for 1.16
There are some changes to the top-level API resources in 1.16. One of
the weired change is about the newly added `TokenRequest` resource
definition where the sole operation defined for it is called
`CreateCoreV1NamespacedServiceAccountToken`. Why we are creating a
resource type named `Potato` and adding an operation called
`CookMushroom`? Sigh.