Make sure responses from operations are sorted
One observation from committing API reference changes to website repo is
that there are many fake changes in the responses section. The response
code is not sorted. This behavior is undesired. It makes the output
unpredictable thus the changes are unclear from a PR commit.

We can sort the responses by the code easily, actually.