Fix API generation for v1.20 (rc.0)
The StorageVersion resource introduced is not a good citizen:

- resource API version named: `internal.apiserver.k8s.io`
- definition path in Swagger: `io.k8s.api.apiserverinternal`
- operation ID defined as: `createInternalApiserver...`

That is, the two key words ("internal" and "apiserver") are in different
order; sometimes they split, other times they are merged.

The following APIs are not RESTful APIs because they are not associated
with any REST resources. These two APIs have to be skipped.

- `getServiceAccountIssuerOpenIDConfiguration` exposed at
  `/.well-known/openid-configuration/`;
- `getServiceAccountIssuerOpenIDKeyset` exposed at `/openid/v1/jwks`.