gen-apidocs: update events api to events.k8s.io
In Kubernetes v1.19, the new Events API events.k8s.io was promoted to
v1. As such it now supersedes the original core Events API. Since this
wasn't defined in this project, the official kubernetes-api doc was
putting the events.k8s.io API under the list of old APIs and the
recommended API for Events was still the core one. To fix that, we need
to define that the events API group is newer than the core one.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>