Fine tune reference generator
This PR updates the templates so that we are outputing references only
for the relevant types. This is to fix a nit where in some outputs
(e.g. kube-proxy-config), irrelevant type definitions got sneaked in.
A better solution would be to export the resource group name into the
configuration file. However, that would be trickier to maintain than
this workaround.