[WIP] Fix component reference generator
There are some command line arguments (e.g.
`default-not-ready-toleration-seconds`) registered for the
`kube-apiserver` as GLOBAL flags. Such kind of flag registration is somehow impacting
the `kube-proxy` command line flags. This PR separates the proxy command
parsing into a separate main file so that API server args are not
impacting the kube-proxy module.