Improve the ConfigAPI reference generator
This PR improves the detection of exposed types by including the
"closest" comment as well because some types are not strictly following
the implicit protocols.
We also relax the constraint for filtering "exported" types. For
example, the (Kube)Config type in the "v1" group should be exposed when
parsing the client-go package.